# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Habit Tracker API",
        "description": "API для гейміфікованого трекера звичок з 3D героями. Підтримує Email/Password, Apple Sign-In та Google Sign-In авторизацію.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/habits/{habit}/log": {
            "post": {
                "tags": [
                    "Habit Logs"
                ],
                "summary": "Log habit completion",
                "description": "Відмітити виконання звички",
                "operationId": "02858012b6bd73782defc084915ddae2",
                "parameters": [
                    {
                        "name": "habit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "completed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-29"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Відчував себе чудово!"
                                    },
                                    "count": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Виконання зареєстровано"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/habits/{habit}/logs": {
            "get": {
                "tags": [
                    "Habit Logs"
                ],
                "summary": "Get habit logs",
                "description": "Історія виконань звички",
                "operationId": "fe30e065beb215331a1a2c79cbc8fba4",
                "parameters": [
                    {
                        "name": "habit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список логів"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/habits/{habit}/stats": {
            "get": {
                "tags": [
                    "Habit Logs"
                ],
                "summary": "Get habit statistics",
                "description": "Статистика виконання звички",
                "operationId": "e13d2946f9cce211f4eff1967a0561d3",
                "parameters": [
                    {
                        "name": "habit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статистика"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/heroes": {
            "get": {
                "tags": [
                    "Heroes"
                ],
                "summary": "Get all heroes",
                "description": "Список усіх доступних героїв",
                "operationId": "418bc30345435e0b6a812ca4b046028e",
                "parameters": [
                    {
                        "name": "is_active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список героїв",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "rarity": {
                                                "type": "string"
                                            },
                                            "unlock_level": {
                                                "type": "integer"
                                            },
                                            "is_premium": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/heroes/{id}": {
            "get": {
                "tags": [
                    "Heroes"
                ],
                "summary": "Get hero details",
                "description": "Деталі героя",
                "operationId": "5ec84c658ea4de745af3035ce1038d60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Деталі героя"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/heroes": {
            "get": {
                "tags": [
                    "User Heroes"
                ],
                "summary": "Get user's heroes",
                "description": "Герої користувача",
                "operationId": "257769e8e12043d38ead634a89adeb93",
                "responses": {
                    "200": {
                        "description": "Список героїв користувача"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/heroes/active": {
            "get": {
                "tags": [
                    "User Heroes"
                ],
                "summary": "Get active hero",
                "description": "Поточний активний герой",
                "operationId": "cdf1dfbb3a2fc2c46956e10e49bb9f3e",
                "responses": {
                    "200": {
                        "description": "Активний герой"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/heroes/{hero}/unlock": {
            "post": {
                "tags": [
                    "User Heroes"
                ],
                "summary": "Unlock hero",
                "description": "Розблокувати героя",
                "operationId": "c113ab4b676af39b5c540a7f62ac7d89",
                "parameters": [
                    {
                        "name": "hero",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Герой розблоковано"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/heroes/{userHero}/activate": {
            "post": {
                "tags": [
                    "User Heroes"
                ],
                "summary": "Activate hero",
                "description": "Активувати героя",
                "operationId": "5bd57884e1f2c9075e23e3d4b5d0074c",
                "parameters": [
                    {
                        "name": "userHero",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Герой активовано"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введіть ваш Bearer token (отриманий при login/register)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}